Changed the project structure, divided the projects into classes.
